requirement.txt
fastapi
uvicorn
aiohttp
httpx
aiomysql
python-dotenv
aiofiles
APScheduler
numpy
scikit-learn
pydantic
loguru
yacs
aiohttp-cors
apscheduler
python-dotenv
aiohttp
apscheduler
aiofiles
filelock
PyYAML
loguru
yacs